fix(plugin): reject control chars in stdio MCP literals
Apply validate_bound_text to stdio argument and environment literals so
CR/LF cannot be smuggled into later spawn. Split the MCP compiler into a
directory module to stay under crate size limits.
